add confirm dialog.

+++ src/app/member-list/member-list.component.ts
@@ -1,19 +1,33 @@
-import {Component, OnInit} from '@angular/core';
+import {Component, OnDestroy, OnInit} from '@angular/core';
 import {MemberService} from '../../services/member.service';
 import {Member} from '../../models/member.model';
+import {ConfirmDialogComponent} from '../../@root/confirm-dialog/confirm-dialog.component';
+import {takeUntil} from 'rxjs/operators';
+import {MatDialog} from '@angular/material/dialog';
+import {Subject} from 'rxjs';
 
 @Component({
   selector: 'app-member-list',
   templateUrl: './member-list.component.html',
   styleUrls: ['./member-list.component.scss']
 })
-export class MemberListComponent implements OnInit {
+export class MemberListComponent implements OnInit, OnDestroy {
+  /** Subject that emits when the component has been destroyed. */
+    // tslint:disable-next-line:variable-name
+  protected _onDestroy = new Subject<void>();
+
   displayedColumns: string[] = ['id', 'cin', 'name', 'type', 'cv', 'createdDate', 'actions'];
   dataSource: Member[] = [];
 
   constructor(
     private memberService: MemberService,
+    private dialog: MatDialog,
   ) {
+  }
+
+  ngOnDestroy(): void {
+    this._onDestroy.next();
+    this._onDestroy.complete();
   }
 
   ngOnInit(): void {
@@ -27,7 +41,18 @@ export class MemberListComponent implements OnInit {
   }
 
   onRemoveAccount(id: any): void {
-    this.memberService.removeMemberById(id).then(() => this.fetchDataSource());
+    const dialogRef = this.dialog.open(ConfirmDialogComponent, {
+      hasBackdrop: true,
+      disableClose: false,
+    });
 
+    dialogRef.componentInstance.confirmButtonColor = 'warn';
+
+    dialogRef.afterClosed().pipe(takeUntil(this._onDestroy)).subscribe(isDeleteConfirmed => {
+      console.log('removing: ', isDeleteConfirmed);
+      if (isDeleteConfirmed) {
+        this.memberService.removeMemberById(id).then(() => this.fetchDataSource());
+      }
+    });
   }
 }
